+++ src/main/java/com/ventionteams/applicationexchange/entity/Lot.java
@@ -1,9 +1,29 @@
 package com.ventionteams.applicationexchange.entity;
 
-import com.ventionteams.applicationexchange.entity.enumeration.*;
-import jakarta.persistence.*;
-import lombok.*;
+import com.ventionteams.applicationexchange.entity.enumeration.Currency;
+import com.ventionteams.applicationexchange.entity.enumeration.LengthUnit;
+import com.ventionteams.applicationexchange.entity.enumeration.LotStatus;
+import com.ventionteams.applicationexchange.entity.enumeration.Packaging;
+import com.ventionteams.applicationexchange.entity.enumeration.Weight;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.EqualsAndHashCode.Include;
+import lombok.NoArgsConstructor;
 
 import java.time.Instant;
 import java.util.ArrayList;
@@ -49,7 +69,7 @@ public class Lot extends AuditingEntity {
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, name = "status")
     private LotStatus status;
-  
+
     @OneToMany(mappedBy = "lot", fetch = FetchType.EAGER)
     private List<Image> images = new ArrayList<>();
 
@@ -71,10 +91,10 @@ public class Lot extends AuditingEntity {
     private Integer bidQuantity;
 
     @Column(nullable = false, name = "total_price")
-    private Double totalPrice;
+    private Long totalPrice;
 
     @Column(nullable = false, name = "start_price")
-    private Double startPrice;
+    private Long startPrice;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, name = "currency")
